+++ Sandswept/Items/Whites/TemporalTransistor.cs
@@ -182,7 +182,8 @@ namespace Sandswept.Items.Whites
 
             if (self.jumpInputReceived && extraJumps > 0 && self.characterMotor && self.characterMotor.jumpCount >= self.characterBody.maxJumpCount && !ignoreRequirements)
             {
-                if (self.isAuthority) {
+                if (self.isAuthority)
+                {
                     self.characterBody.SetBuffCountSynced(extraJump.buffIndex, extraJumps - 1);
                 }
 
@@ -229,7 +230,7 @@ namespace Sandswept.Items.Whites
         public override ItemDisplayRuleDict CreateItemDisplayRules()
         {
             return new ItemDisplayRuleDict();
-            
+
             var itemDisplay = SetUpFollowerIDRS(0.72f, 132f, true, 30f, false, 0f, true, 20f);
 
             return new ItemDisplayRuleDict(new ItemDisplayRule()